+++ confidential-identities/build.gradle
@@ -1,5 +1,4 @@
-// This contains the SwapIdentitiesFlow which can be used for exchanging confidential identities as part of a flow.
-// TODO: Merge this into core: the original plan was to develop it independently but in practice it's too widely used to break compatibility now, as finance uses it.
+// This has been merged into core/node
 apply plugin: 'kotlin'
 apply plugin: CanonicalizerPlugin
 apply plugin: 'net.corda.plugins.publish-utils'
@@ -11,20 +10,6 @@ description 'Corda Experimental Confidential Identities'
 
 dependencies {
     cordaCompile project(':core')
-    cordaCompile (group: 'net.corda', name: 'corda-node', version: '5.0-SNAPSHOT')
-
-    testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
-    testCompile "junit:junit:$junit_version"
-
-    // Guava: Google test library (collections test suite)
-    testCompile "com.google.guava:guava-testlib:$guava_version"
-
-    // Bring in the MockNode infrastructure for writing protocol unit tests.
-    testCompile project(":node-driver")
-
-    // AssertJ: for fluent assertions for testing
-    testCompile "org.assertj:assertj-core:$assertj_version"
-    testCompile (group: 'net.corda', name: 'corda-node', version: '5.0-SNAPSHOT')
 }
 
 jar {
