+++ core/src/main/kotlin/net/corda/core/identity/IdentityUtils.kt
@@ -2,8 +2,10 @@
 
 package net.corda.core.identity
 
+import net.corda.core.crypto.DigitalSignature
 import net.corda.core.internal.toMultiMap
 import net.corda.core.node.ServiceHub
+import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import java.security.PublicKey
 
@@ -77,3 +79,12 @@ fun <T> excludeHostNode(serviceHub: ServiceHub, map: Map<Party, T>): Map<Party, 
  * @return a new copy of the map, with the well known [Party] for the notary removed.
  */
 fun <T> excludeNotary(map: Map<Party, T>, stx: SignedTransaction): Map<Party, T> = map.filterKeys { it != stx.notary }
+
+@CordaSerializable
+data class KeyToPartyMapping(val key: PublicKey, val party: Party)
+
+@CordaSerializable
+data class SignedKeyToPartyMapping(
+        val mapping: KeyToPartyMapping,
+        val signature: DigitalSignature.WithKey
+)
