+++ core/src/test/kotlin/net/corda/core/transactions/LedgerTransactionQueryTests.kt
@@ -9,11 +9,10 @@ import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.identity.Party
 import net.corda.core.node.NotaryInfo
-import net.corda.node.services.api.IdentityServiceInternal
+import net.corda.core.node.services.IdentityService
 import net.corda.testing.common.internal.testNetworkParameters
 import net.corda.testing.contracts.DummyContract
 import net.corda.testing.core.*
-import net.corda.testing.internal.rigorousMock
 import net.corda.testing.node.MockServices
 import org.junit.Before
 import org.junit.Rule
@@ -35,7 +34,7 @@ class LedgerTransactionQueryTests {
     private val services = MockServices(
             listOf("net.corda.testing.contracts"),
             TestIdentity(CordaX500Name("MegaCorp", "London", "GB"), keyPair),
-            mock<IdentityServiceInternal>().also {
+            mock<IdentityService>().also {
                 doReturn(null).whenever(it).partyFromKey(keyPair.public)
             },
             testNetworkParameters(notaries = listOf(NotaryInfo(DUMMY_NOTARY, true))),
