+++ finance/workflows/build.gradle
@@ -28,13 +28,13 @@ dependencies {
     // cordapp project(':finance:workflows')
     // cordapp project(':finance:contracts')
     cordaCompile project(':core')
-    cordaCompile project(':confidential-identities')
-    
+    cordaCompile project(':node')
+
     cordapp project(':finance:contracts')
 
     // For JSON
     compile "com.fasterxml.jackson.core:jackson-databind:${jackson_version}"
-    
+
     testCompile project(':test-utils')
     testCompile project(path: ':core', configuration: 'testArtifacts')
     testCompile "junit:junit:$junit_version"
