+++ finance/workflows/src/main/kotlin/net/corda/finance/flows/CashPaymentFlow.kt
@@ -1,7 +1,7 @@
 package net.corda.finance.flows
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.confidential.SwapIdentitiesFlow
+import net.corda.confidential.identities.SwapIdentitiesFlow
 import net.corda.core.contracts.Amount
 import net.corda.core.contracts.InsufficientBalanceException
 import net.corda.core.flows.*
