+++ finance/workflows/src/main/kotlin/net/corda/finance/flows/TwoPartyTradeFlow.kt
@@ -1,7 +1,7 @@
 package net.corda.finance.flows
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.confidential.IdentitySyncFlow
+import net.corda.confidential.identities.IdentitySyncFlow
 import net.corda.core.contracts.*
 import net.corda.core.flows.*
 import net.corda.core.identity.Party
