+++ node/src/main/kotlin/net/corda/confidential/identities/SwapIdentitiesFlow.kt
@@ -1,4 +1,4 @@
-package net.corda.confidential
+package net.corda.confidential.identities
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.CordaInternal
