+++ node/src/main/kotlin/net/corda/node/internal/SwapIdentitiesHandler.kt
@@ -1,10 +1,9 @@
 package net.corda.node.internal
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.confidential.SwapIdentitiesFlow
+import net.corda.confidential.identities.SwapIdentitiesFlow
 import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.FlowSession
-import net.corda.core.flows.InitiatedBy
 import net.corda.core.internal.warnOnce
 
 /**
