+++ node/src/main/kotlin/net/corda/node/services/identity/PersistentIdentityService.kt
@@ -2,21 +2,25 @@ package net.corda.node.services.identity
 
 import net.corda.core.crypto.SecureHash
 import net.corda.core.identity.*
+import net.corda.core.internal.CertRole
 import net.corda.core.internal.NamedCacheFactory
 import net.corda.core.internal.hash
+import net.corda.core.node.services.IdentityService
 import net.corda.core.node.services.UnknownAnonymousPartyException
+import net.corda.core.node.services.x500Matches
 import net.corda.core.serialization.SingletonSerializeAsToken
 import net.corda.core.utilities.MAX_HASH_HEX_SIZE
 import net.corda.core.utilities.contextLogger
 import net.corda.core.utilities.debug
 import net.corda.node.internal.schemas.NodeInfoSchemaV1
-import net.corda.node.services.api.IdentityServiceInternal
 import net.corda.node.utilities.AppendOnlyPersistentMap
 import net.corda.nodeapi.internal.crypto.X509CertificateFactory
+import net.corda.nodeapi.internal.crypto.X509Utilities
 import net.corda.nodeapi.internal.crypto.x509Certificates
 import net.corda.nodeapi.internal.persistence.CordaPersistence
 import net.corda.nodeapi.internal.persistence.NODE_DATABASE_PREFIX
 import org.apache.commons.lang.ArrayUtils.EMPTY_BYTE_ARRAY
+import java.lang.IllegalArgumentException
 import java.security.InvalidAlgorithmParameterException
 import java.security.PublicKey
 import java.security.cert.*
@@ -31,7 +35,8 @@ import javax.persistence.Lob
  * cached for efficient lookup.
  */
 @ThreadSafe
-class PersistentIdentityService(cacheFactory: NamedCacheFactory) : SingletonSerializeAsToken(), IdentityServiceInternal {
+class PersistentIdentityService(cacheFactory: NamedCacheFactory) : SingletonSerializeAsToken(), IdentityService {
+
     companion object {
         private val log = contextLogger()
 
@@ -167,11 +172,44 @@ class PersistentIdentityService(cacheFactory: NamedCacheFactory) : SingletonSeri
     }
 
     @Throws(CertificateExpiredException::class, CertificateNotYetValidException::class, InvalidAlgorithmParameterException::class)
-    override fun verifyAndRegisterIdentity(identity: PartyAndCertificate, isNewRandomIdentity: Boolean): PartyAndCertificate? {
+    private fun verifyAndRegisterIdentity(identity: PartyAndCertificate, isNewRandomIdentity: Boolean): PartyAndCertificate? {
         return database.transaction {
             verifyAndRegisterIdentity(trustAnchor, identity, isNewRandomIdentity)
         }
     }
+
+    /**
+     * Verifies that an identity is valid.
+     *
+     * @param trustAnchor The trust anchor that will verify the identity's validity
+     * @param identity The identity to verify
+     * @param isNewRandomIdentity true if the identity will not have been registered before (e.g. because it is randomly generated by ourselves).
+     */
+    @Throws(CertificateExpiredException::class, CertificateNotYetValidException::class, InvalidAlgorithmParameterException::class)
+    private fun verifyAndRegisterIdentity(trustAnchor: TrustAnchor, identity: PartyAndCertificate, isNewRandomIdentity: Boolean = false): PartyAndCertificate? {
+        // Validate the chain first, before we do anything clever with it
+        val identityCertChain = identity.certPath.x509Certificates
+        try {
+            identity.verify(trustAnchor)
+        } catch (e: CertPathValidatorException) {
+            log.warn("Certificate validation failed for ${identity.name} against trusted root ${trustAnchor.trustedCert.subjectX500Principal}.")
+            log.warn("Certificate path :")
+            identityCertChain.reversed().forEachIndexed { index, certificate ->
+                val space = (0 until index).joinToString("") { "   " }
+                log.warn("$space${certificate.subjectX500Principal}")
+            }
+            throw e
+        }
+        // Ensure we record the first identity of the same name, first
+        val wellKnownCert = identityCertChain.single { CertRole.extract(it)?.isWellKnown ?: false }
+        if (wellKnownCert != identity.certificate && !isNewRandomIdentity) {
+            val idx = identityCertChain.lastIndexOf(wellKnownCert)
+            val firstPath = X509Utilities.buildCertPath(identityCertChain.slice(idx until identityCertChain.size))
+            verifyAndRegisterIdentity(trustAnchor, PartyAndCertificate(firstPath))
+        }
+        return registerIdentity(identity, isNewRandomIdentity)
+    }
+
 
     override fun registerIdentity(identity: PartyAndCertificate, isNewRandomIdentity: Boolean): PartyAndCertificate? {
         log.debug { "Registering identity $identity" }
@@ -190,10 +228,6 @@ class PersistentIdentityService(cacheFactory: NamedCacheFactory) : SingletonSeri
             val parentId = mapToKey(identityCertChain[1].publicKey)
             keyToPartyAndCert[parentId]
         }
-    }
-
-    fun registerConfidentialIdentity(key: PublicKey) {
-        log.debug { "Registering confidential identity with key $key" }
     }
 
     @Deprecated("This should be removed. But probably won't be. So don't use. ")
@@ -268,18 +302,30 @@ class PersistentIdentityService(cacheFactory: NamedCacheFactory) : SingletonSeri
         return keys.filter { certificateFromKey(it)?.name in ourNames }
     }
 
-    override fun registerIdentityMapping(identity: Party, key: PublicKey): Boolean {
-        var willRegisterNewMapping: Boolean = true
+    override fun registerConfidentialIdentity(keyMapping: SignedKeyToPartyMapping, nodeParty: Party): Boolean {
+        val k = keyMapping.mapping.key
+        val p = keyMapping.mapping.party
+        val sig = keyMapping.signature
+
+        if (p != nodeParty) {
+            throw IllegalArgumentException("Something something something")
+        }
+
+        if (sig.by != nodeParty.owningKey) {
+            throw IllegalArgumentException("Something somethign seomthien ghrijei")
+        }
+
+        var willRegisterNewMapping = true
 
         database.transaction {
-            val existingEntryForKey = keyToParty[key.hash]
+            val existingEntryForKey = keyToParty[k.hash]
             if (existingEntryForKey == null) {
-                log.info("Linking: ${key.hash} to ${identity.name}")
-                keyToParty[key.hash] = identity.name
+                log.info("Linking: ${k.hash} to ${p.name}")
+                keyToParty[k.hash] = p.name
             } else {
-                log.info("An existing entry for ${key.hash} already exists.")
-                if (identity.name != keyToParty[key.hash]) {
-                    log.error("The public key ${key.hash} is already assigned to a party.")
+                log.info("An existing entry for ${k.hash} already exists.")
+                if (p.name != keyToParty[k.hash]) {
+                    log.error("The public key ${k.hash} is already assigned to a party.")
                 }
                 willRegisterNewMapping = false
             }
