+++ node/src/main/kotlin/net/corda/node/services/keys/KMSUtils.kt
@@ -5,7 +5,6 @@ import net.corda.core.identity.PartyAndCertificate
 import net.corda.core.internal.CertRole
 import net.corda.core.node.services.IdentityService
 import net.corda.core.utilities.days
-import net.corda.node.services.api.IdentityServiceInternal
 import net.corda.nodeapi.internal.crypto.CertificateType
 import net.corda.nodeapi.internal.crypto.ContentSignerBuilder
 import net.corda.nodeapi.internal.crypto.X509Utilities
@@ -45,11 +44,8 @@ fun freshCertificate(identityService: IdentityService,
             window)
     val ourCertPath = X509Utilities.buildCertPath(ourCertificate, issuer.certPath.x509Certificates)
     val anonymisedIdentity = PartyAndCertificate(ourCertPath)
-    if (identityService is IdentityServiceInternal) {
-        identityService.justVerifyAndRegisterIdentity(anonymisedIdentity, true)
-    } else {
+        //TODO fix this?
         identityService.verifyAndRegisterIdentity(anonymisedIdentity)
-    }
     return anonymisedIdentity
 }
